admin

+++ app/Http/Controllers/LayananController.php
@@ -3,10 +3,12 @@
 namespace App\Http\Controllers;
 
 use App\Models\Layanan;
+use App\Models\Gambar;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
 // use Cloudinary\Cloudinary;
 use CloudinaryLabs\CloudinaryLaravel\Facades\Cloudinary;
+use Illuminate\Support\Facades\DB;
 
 class LayananController extends Controller
 {
@@ -15,9 +17,15 @@ class LayananController extends Controller
      */
     public function index()
     {
+        // $data = DB::table('layanans')
+        // ->join('gambars', 'gambars.layananId', '=', 'layanans.id')
+        // ->select('layanans.*', 'gambars.* as gambar')
+        // ->get();
+        // dd($data);
+        // dd(Layanan::with('gambar')->get());
         return Inertia::render('Admin/Layanan/Index', [
             'title' => 'Layanan',
-            'layanan' => Layanan::all(),
+            'layanan' => Layanan::with('gambar')->get(),
         ]);
     }
 
@@ -42,16 +50,29 @@ class LayananController extends Controller
      */
     public function store(Request $request)
     {
-        Layanan::create([
-            'namaLayanan' => $request->namaLayanan,
-            'deskripsi' => $request->deskripsi,
-            'hrgTerrendah' => $request->hrgTerrendah,
-            'hrgTertinggi' => $request->hrgTertinggi,
-            'imgName' => $request->imgName,
-            'imgUrl' => $request->imgUrl,
-            'videoName' => $request->videoName,
-            'videoUrl' => $request->videoUrl,
+        // dd($request->image);
+        $layanan = Layanan::create([
+            'namaLayanan' => $request->input("values.namaLayanan"),
+            'deskripsi' => $request->input("values.deskripsi"),
+            'hrgTerrendah' => $request->input('values.hrgTerrendah'),
+            'hrgTertinggi' => $request->input("values.hrgTertinggi"),
+            // 'imgName' => $request->imgName,
+            // 'imgUrl' => $request->imgUrl,
+            'videoName' => $request->input("values.videoName"),
+            'videoUrl' => $request->input("values.videoUrl"),
         ]);
+
+        // dd($layanan);
+
+        foreach($request->input('image') as $d) {
+            $g = [
+                'layananId'  => $layanan->id,
+                'imgName' =>  $d['public_id'],
+                'imgUrl' => $d['url'],
+            ];
+            Gambar::create($g);
+
+        }
 
         // return response()->json(['data' => 'Berhasil menambah']);
         return redirect('/admin/layanan')->with(['data' => 'Berhasil menambah']);
@@ -70,7 +91,21 @@ class LayananController extends Controller
      */
     public function edit(Layanan $layanan)
     {
-        //
+        // $d = DB::table('layanans')
+        // ->join('gambars', 'gambars.layananId', '=', 'layanans.id')
+        // ->where('layanans.id', $layanan->id)
+        // ->select('layanans.*', 'gambars.* as gambar')
+        // ->get();
+        $gambar = Gambar::where('layananId', $layanan->id)->get();
+        // dd($gambar);
+        $layanan = $layanan;
+
+
+        return Inertia::render('Admin/Layanan/Edit', [
+            'title' => 'Edit Layanan',
+            'layanan' => $layanan,
+            'gambar' => $gambar,
+        ]);
     }
 
     /**
@@ -86,8 +121,14 @@ class LayananController extends Controller
      */
     public function destroy(Layanan $layanan)
     {
-        Cloudinary::destroy($layanan->imgName);
-        Cloudinary::destroy($layanan->videoName, ["resource_type" => "video"]);
+        $gambar = Gambar::where('id', $layanan->id)->get();
+        foreach ($gambar as $key => $g) {
+            Cloudinary::destroy($g->imgName);
+        }
+        if($layanan->videoName != '') {
+            Cloudinary::destroy($layanan->videoName, ["resource_type" => "video"]);
+        }
+
         $layanan->delete();
         // return response()->json(['data' => 'Berhasil menghapus']);
         return redirect('/admin/layanan')->with(['data' => 'Berhasil menghapus']);
